Setup guardian project structure
