+++ apps/guardian/.gitignore
@@ -22,3 +22,8 @@ dist-ssr
 *.njsproj
 *.sln
 *.sw?
+
+# Cargo project
+# will have compiled files and executables
+src-tauri/target
+
