--- apps/guardian/src-tauri/.gitignore
@@ -1,4 +1,0 @@
-# Generated by Cargo
-# will have compiled files and executables
-/target/
-
