+++ apps/guardian/src-tauri/build.rs
@@ -1,3 +1,1 @@
-fn main() {
-    tauri_build::build()
-}
+fn main() { tauri_build::build() }
