+++ apps/guardian/src-tauri/src/main.rs
@@ -3,9 +3,7 @@
 
 // Learn more about Tauri commands at https://tauri.app/v1/guides/features/command
 #[tauri::command]
-fn greet(name: &str) -> String {
-    format!("Hello, {}! You've been greeted from Rust!", name)
-}
+fn greet(name: &str) -> String { format!("Hello, {}! You've been greeted from Rust!", name) }
 
 fn main() {
     tauri::Builder::default()
