+++ apps/guardian/src/main.tsx
@@ -1,10 +1,11 @@
-import React from "react";
-import ReactDOM from "react-dom/client";
-import App from "./App";
-import "./styles.css";
+import React from 'react';
+import { createRoot } from 'react-dom/client';
 
-ReactDOM.createRoot(document.getElementById("root") as HTMLElement).render(
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>,
+import Root from '@shell/Root';
+import './scss/index.scss';
+
+createRoot(document.getElementById('root')!).render(
+    <React.StrictMode>
+        <Root />
+    </React.StrictMode>,
 );
