+++ apps/guardian/src/pages/Home/Home.tsx
@@ -1,0 +1,52 @@
+import { useState } from 'react';
+import reactLogo from '../../assets/react.svg';
+import { invoke } from '@tauri-apps/api/tauri';
+
+function App() {
+    const [greetMsg, setGreetMsg] = useState('');
+    const [name, setName] = useState('');
+
+    async function greet() {
+        // Learn more about Tauri commands at https://tauri.app/v1/guides/features/command
+        setGreetMsg(await invoke<string>('greet', { name }));
+    }
+
+    return (
+        <div className="container">
+            <h1>Welcome to Tauri!</h1>
+
+            <div className="row">
+                <a href="https://vitejs.dev" target="_blank" rel="noreferrer">
+                    <img src="/vite.svg" className="logo vite" alt="Vite logo" />
+                </a>
+                <a href="https://tauri.app" target="_blank" rel="noreferrer">
+                    <img src="/tauri.svg" className="logo tauri" alt="Tauri logo" />
+                </a>
+                <a href="https://reactjs.org" target="_blank" rel="noreferrer">
+                    <img src={reactLogo} className="logo react" alt="React logo" />
+                </a>
+            </div>
+
+            <p>Click on the Tauri, Vite, and React logos to learn more.</p>
+
+            <form
+                className="row"
+                onSubmit={(e) => {
+                    e.preventDefault();
+                    void greet();
+                }}
+            >
+                <input
+                    id="greet-input"
+                    onChange={(e) => setName(e.currentTarget.value)}
+                    placeholder="Enter a name..."
+                />
+                <button type="submit">Greet</button>
+            </form>
+
+            <p>{greetMsg}</p>
+        </div>
+    );
+}
+
+export default App;
