+++ apps/guardian/src/pages/Home/_Home.scss
@@ -1,0 +1,3 @@
+.home {
+    display: block;
+}
