+++ apps/guardian/src/pages/_bundle.scss
@@ -1,0 +1,1 @@
+@import './Home/Home';
