+++ apps/guardian/src/scss/_components.scss
@@ -1,0 +1,2 @@
+// scss files for application components declared below.
+@import '../pages/bundle';
