+++ apps/guardian/src/scss/_layout.scss
@@ -1,0 +1,2 @@
+// Project layout - add more imports from the `layout` folder as they are used.
+@import 'layout/root';
