+++ apps/guardian/src/scss/index.scss
@@ -1,0 +1,4 @@
+// @import 'ccd-bootstrap/scss/index.scss';
+
+@import 'layout';
+@import 'components';
