+++ apps/guardian/src/shell/router.tsx
@@ -1,0 +1,9 @@
+import { createBrowserRouter } from 'react-router-dom';
+import Home from '@pages/Home';
+
+export const router = createBrowserRouter([
+    {
+        path: '/',
+        element: <Home />,
+    },
+]);
